remove unused code, move some funcs around
