+++ protocol/bob/swap_state.go
@@ -268,129 +268,6 @@ func (s *swapState) tryClaim() (ethcommon.Hash, error) {
 	return s.claimFunds()
 }
 
-// HandleProtocolMessage is called by the network to handle an incoming message.
-// If the message received is not the expected type for the point in the protocol we're at,
-// this function will return an error.
-func (s *swapState) HandleProtocolMessage(msg net.Message) (net.Message, bool, error) {
-	s.Lock()
-	defer s.Unlock()
-
-	if s.ctx.Err() != nil {
-		return nil, true, fmt.Errorf("protocol exited: %w", s.ctx.Err())
-	}
-
-	if err := s.checkMessageType(msg); err != nil {
-		return nil, true, err
-	}
-
-	switch msg := msg.(type) {
-	case *net.SendKeysMessage:
-		if err := s.handleSendKeysMessage(msg); err != nil {
-			return nil, true, err
-		}
-
-		// we initiated, so we're now waiting for Alice to deploy the contract.
-		return nil, false, nil
-	case *net.NotifyContractDeployed:
-		if msg.Address == "" {
-			return nil, true, errMissingAddress
-		}
-
-		log.Infof("got Swap contract address! address=%s", msg.Address)
-
-		if err := s.setContract(ethcommon.HexToAddress(msg.Address)); err != nil {
-			return nil, true, fmt.Errorf("failed to instantiate contract instance: %w", err)
-		}
-
-		fp := fmt.Sprintf("%s/%d/contractaddress", s.bob.basepath, s.id)
-		if err := common.WriteContractAddressToFile(fp, msg.Address); err != nil {
-			return nil, true, fmt.Errorf("failed to write contract address to file: %w", err)
-		}
-
-		if err := s.checkContract(); err != nil {
-			return nil, true, err
-		}
-
-		addrAB, err := s.lockFunds(s.providesAmount)
-		if err != nil {
-			return nil, true, fmt.Errorf("failed to lock funds: %w", err)
-		}
-
-		out := &net.NotifyXMRLock{
-			Address: string(addrAB),
-		}
-
-		// set t0 and t1
-		if err := s.setTimeouts(); err != nil {
-			return nil, true, err
-		}
-
-		go func() {
-			until := time.Until(s.t0)
-
-			log.Debugf("time until t0: %vs", until.Seconds())
-
-			select {
-			case <-s.ctx.Done():
-				return
-			case <-time.After(until + time.Second):
-				// we can now call Claim()
-				txHash, err := s.claimFunds()
-				if err != nil {
-					log.Errorf("failed to claim: err=%s", err)
-					// TODO: retry claim, depending on error
-					return
-				}
-
-				log.Debug("funds claimed!")
-				s.completed = true
-
-				// send *net.NotifyClaimed
-				if err := s.bob.net.SendSwapMessage(&net.NotifyClaimed{
-					TxHash: txHash.String(),
-				}); err != nil {
-					log.Errorf("failed to send NotifyClaimed message: err=%s", err)
-				}
-			case <-s.readyCh:
-				return
-			}
-		}()
-
-		s.nextExpectedMessage = &net.NotifyReady{}
-		return out, false, nil
-	case *net.NotifyReady:
-		log.Debug("Alice called Ready(), attempting to claim funds...")
-		close(s.readyCh)
-
-		// contract ready, let's claim our ether
-		txHash, err := s.claimFunds()
-		if err != nil {
-			return nil, true, fmt.Errorf("failed to redeem ether: %w", err)
-		}
-
-		log.Debug("funds claimed!!")
-		out := &net.NotifyClaimed{
-			TxHash: txHash.String(),
-		}
-
-		s.completed = true
-		return out, true, nil
-	case *net.NotifyRefund:
-		// generate monero wallet, regaining control over locked funds
-		addr, err := s.handleRefund(msg.TxHash)
-		if err != nil {
-			return nil, false, err
-		}
-
-		s.completed = true
-		s.refunded = true
-		log.Infof("regained control over monero account %s", addr)
-		return nil, true, nil
-	default:
-		return nil, true, errors.New("unexpected message type")
-	}
-}
-
 // generateKeys generates Bob's spend and view keys (s_b, v_b)
 // It returns Bob's public spend key and his private view key, so that Alice can see
 // if the funds are locked.
@@ -513,60 +390,6 @@ func (s *swapState) checkContract() error {
 	skTheirs := s.aliceSecp256K1PublicKey.Keccak256()
 	if !bytes.Equal(pkRefund[:], skOurs[:]) {
 		return fmt.Errorf("contract claim key is not expected: got 0x%x, expected 0x%x", pkRefund, skTheirs)
-	}
-
-	return nil
-}
-
-func (s *swapState) handleSendKeysMessage(msg *net.SendKeysMessage) error {
-	if msg.PublicSpendKey == "" || msg.PublicViewKey == "" {
-		return errMissingKeys
-	}
-
-	log.Debug("got Alice's public keys")
-
-	kp, err := mcrypto.NewPublicKeyPairFromHex(msg.PublicSpendKey, msg.PublicViewKey)
-	if err != nil {
-		return fmt.Errorf("failed to generate Alice's public keys: %w", err)
-	}
-
-	// verify counterparty's DLEq proof and ensure the resulting secp256k1 key is correct
-	secp256k1Pub, err := pcommon.VerifyKeysAndProof(msg.DLEqProof, msg.Secp256k1PublicKey)
-	if err != nil {
-		return err
-	}
-
-	s.setAlicePublicKeys(kp, secp256k1Pub)
-	s.nextExpectedMessage = &net.NotifyContractDeployed{}
-	return nil
-}
-
-func (s *swapState) handleRefund(txHash string) (mcrypto.Address, error) {
-	receipt, err := s.bob.ethClient.TransactionReceipt(s.ctx, ethcommon.HexToHash(txHash))
-	if err != nil {
-		return "", err
-	}
-
-	if len(receipt.Logs) == 0 {
-		return "", errors.New("claim transaction has no logs")
-	}
-
-	sa, err := swap.GetSecretFromLog(receipt.Logs[0], "Refunded")
-	if err != nil {
-		return "", err
-	}
-
-	return s.reclaimMonero(sa)
-}
-
-func (s *swapState) checkMessageType(msg net.Message) error {
-	// Alice might refund anytime before t0 or after t1, so we should allow this.
-	if _, ok := msg.(*net.NotifyRefund); ok {
-		return nil
-	}
-
-	if msg.Type() != s.nextExpectedMessage.Type() {
-		return errors.New("received unexpected message")
 	}
 
 	return nil
